Implemented sending user_data.json to connected clients, so we can add things like API keys for different visualizations, or configuration information for how we want things to look, differently from what's built in.
Added a new visualization based on Travelling Wavefronts, from https://www.cake23.de/traveling-wavefronts-lit-up.html.  Completely refactored and nearly re-wrote the code, also changed many of the visualization parameters to make it a lot more interesting.  We can still improve it, but it already looks so good.

Changed our background organization code to make it even easier to swap things out.  Streamlined the requirements for how to add visualizations to the project, making them even more drop-in.

Updated the Readme.